backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.4
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-toolbelt==0.0.1
djangorestframework==3.12.2
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
heroku==0.1.4
httpie==2.4.0
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
Pillow==8.1.0
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.7.4
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==1.5
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
ruamel.yaml==0.16.12
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
